preserve dtype

+++ cirq/linalg/distance_measures.py
@@ -16,14 +16,12 @@
 import numpy as np
 import scipy
 
-from cirq import linalg
-
 
 def _sqrt_positive_semidefinite_matrix(mat: np.ndarray) -> np.ndarray:
     """Square root of a positive semidefinite matrix."""
     eigs, vecs = scipy.linalg.eigh(mat)
     # Zero out small negative entries
-    eigs = np.maximum(eigs, np.zeros(eigs.shape))
+    eigs = np.maximum(eigs, np.zeros(eigs.shape, dtype=eigs.dtype))
     return vecs @ np.diag(np.sqrt(eigs)) @ vecs.T.conj()
